+++ lib/hutch/cli.rb
@@ -97,7 +97,7 @@ module Hutch
           Hutch::Config.mq_port = port
         end
 
-        opts.on('--mq-tls', 'Use TLS for the AMQP connection') do |tls|
+        opts.on("-t", "--[no-]mq-tls", 'Use TLS for the AMQP connection') do |tls|
           Hutch::Config.mq_tls = tls
         end
 
@@ -138,7 +138,7 @@ module Hutch
           Hutch::Config.mq_api_port = port
         end
 
-        opts.on('--mq-api-ssl', 'Use SSL for the RabbitMQ API') do |api_ssl|
+        opts.on("-s", "--[no-]mq-api-ssl", 'Use SSL for the RabbitMQ API') do |api_ssl|
           Hutch::Config.mq_api_ssl = api_ssl
         end