+++ spec/spec_helper.rb
@@ -1,3 +1,14 @@
+require 'simplecov'
+if ENV['TRAVIS']
+  require 'coveralls'
+  SimpleCov.formatter = Coveralls::SimpleCov::Formatter
+end
+
+SimpleCov.start do
+  add_filter '/spec/'
+  add_filter '/.bundle/'
+end
+
 require 'raven'
 require 'hutch'
 require 'logger'